refactor: extract verify to separate crate
