+++ crates/forge/Cargo.toml
@@ -45,6 +45,7 @@ yansi = "0.5"
 # bin
 forge-doc.workspace = true
 forge-fmt.workspace = true
+forge-verify.workspace = true
 foundry-cli.workspace = true
 foundry-debugger.workspace = true
