+++ crates/forge/bin/cmd/mod.rs
@@ -56,12 +56,10 @@ pub mod inspect;
 pub mod install;
 pub mod remappings;
 pub mod remove;
-pub mod retry;
 pub mod script;
 pub mod selectors;
 pub mod snapshot;
 pub mod test;
 pub mod tree;
 pub mod update;
-pub mod verify;
 pub mod watch;
