+++ crates/verify/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "forge-verify"
+description = "Contract verification tools"
+
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
+foundry-config.workspace = true
+foundry-cli.workspace = true
+foundry-common.workspace = true
+foundry-evm.workspace = true
+
+serde_json.workspace = true
+hex.workspace = true
+alloy-json-abi.workspace = true
+alloy-primitives.workspace = true
+serde.workspace = true
+eyre.workspace = true
+ethers-providers.workspace = true
+tracing.workspace = true
+foundry-compilers = { workspace = true, features = ["full"] }
+foundry-block-explorers = { workspace = true, features = ["foundry-compilers"] }
+
+clap = { version = "4", features = ["derive", "env", "unicode", "wrap_help"] }
+reqwest = { version = "0.11", default-features = false, features = ["json"] }
+async-trait = "0.1"
+futures = "0.3"
+semver = "1"
+regex = { version = "1", default-features = false }
+once_cell = "1"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["macros"] }
+foundry-test-utils.workspace = true
+tempfile = "3"
