+++ crates/verify/src/lib.rs
@@ -1,4 +1,6 @@
-use super::retry::RetryArgs;
+#[macro_use]
+extern crate tracing;
+
 use alloy_primitives::Address;
 use clap::{Parser, ValueHint};
 use eyre::Result;
@@ -19,7 +21,10 @@ use etherscan::EtherscanVerificationProvider;
 pub mod provider;
 use provider::VerificationProvider;
 
+pub mod retry;
 mod sourcify;
+
+pub use retry::RetryArgs;
 
 /// Verification provider arguments
 #[derive(Clone, Debug, Parser)]
